Extract Requests to separate files and namespace
